refactor: implement CodeHighlight component and update styles in Tailwind CSS configuration
