+++ package.json
@@ -50,7 +50,7 @@
     "zustand": "^5.0.5"
   },
   "devDependencies": {
-    "@cloudflare/vite-plugin": "^1.5.0",
+    "@cloudflare/vite-plugin": "^1.6.0",
     "@epic-web/config": "^1.21.0",
     "@react-router/dev": "7.6.1",
     "@tailwindcss/typography": "^0.5.16",
